post, comment crud

+++ Models/Post.cs
@@ -13,8 +13,8 @@ namespace KozoskodoAPI.Models
         [Column(TypeName = "int(11)")]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int Id { get; set; }
-        public byte? Loves { get; set; }
-        public byte? DisLoves { get; set; }
+        public uint? Loves { get; set; }
+        public uint? DisLoves { get; set; }
         public DateTime DateOfPost { get; set; }
         [StringLength(500)]
         public string? PostContent { get; set; }
